[backend] Outras Entidades

+++ backend/src/modules/periodos/services/deletePeriodo.service.ts
@@ -1,5 +1,8 @@
 import { Injectable } from '@nestjs/common';
 
+import { AppError } from '@shared/errors/AppError';
+
+import { periodoErrors } from '../errors/periodo.errors';
 import { PeriodosRepository } from '../repositories/periodos.repository';
 import { CommonPeriodoService } from './commonPeriodo.service';
 
@@ -13,7 +16,12 @@ export class DeletePeriodoService {
   async execute(id: string) {
     const periodo = await this.commonPeriodoService.getPeriodo({
       id,
+      relations: ['vinculos'],
     });
+
+    if (periodo.vinculos.length > 0) {
+      throw new AppError(periodoErrors.deleteWithVinculos);
+    }
 
     await this.periodosRepository.delete(periodo);
   }
